imap: move over imapclient.StoreFlags
