+++ store.go
@@ -1,0 +1,17 @@
+package imap
+
+// StoreFlagsOp is a flag operation: set, add or delete.
+type StoreFlagsOp int
+
+const (
+	StoreFlagsSet StoreFlagsOp = iota
+	StoreFlagsAdd
+	StoreFlagsDel
+)
+
+// StoreFlags alters message flags.
+type StoreFlags struct {
+	Op     StoreFlagsOp
+	Silent bool
+	Flags  []Flag
+}
